Extract the scheduler to lib/scheduler.js
Verbatim move of three server.js spans: routineState + its persistence
(loadRoutineState/saveRoutineState/recordRoutineRun), and the scheduler
proper (startScheduler, getNextRun, executeRoutine, the routine
broadcast). Mechanical conversions, each count-asserted at extraction:
2 cwd sites read getWorkspace() at use time, 1 wss.clients read goes
through the injected getWssClients accessor, 4 Claude spawn-plumbing
calls (spawnClaude/getBareArgs/modelArgs/getSpawnEnv) go through wired
deps until spawn plumbing gets its own extraction. Direct lib requires
(no wiring, same names): rundockDir, recordEvent, buildSystemPrompt,
getCodexAppServer/waitForCodexReady/readAgentInstructions,
discoverAgents.

routineState is module-owned and exported BY IDENTITY; the root
re-exports it (plus the six functions) unchanged for the test surface,
so every existing scheduler test runs against the moved code with zero
retargets. The root's agentsDiscovery.setRoutineState wiring keeps
working this commit (it now passes the module's object); the next
commit retires it in favour of a direct lib-to-lib require.

The two glue destructures only the moved code used (getCodexAppServer,
waitForCodexReady) leave the root's require list.

server.js: 5,573 -> 5,404 lines. Suite 1458 green.
