+++ server.js
@@ -488,14 +488,15 @@ function discoverWorkspaces() {
   return candidates;
 }
 
-// ===== ROUTINE STATE =====
-// In-memory view of routine run state, persisted to .rundock/routine-state.json
-// so a server restart cannot re-fire a routine that already ran in its window
-// (the desktop quit-and-reopen pattern). The file is workspace-scoped like the
-// other .rundock stores; loadRoutineState() runs at startup and on every
-// workspace switch.
-
-const routineState = {}; // { routineKey: { lastRun, status, duration } }
+// ===== ROUTINE STATE + SCHEDULER =====
+// Routine state, its persistence, and the scheduler (startScheduler,
+// getNextRun, executeRoutine) live in lib/scheduler.js. routineState comes
+// back BY IDENTITY for the test re-exports below.
+const schedulerLib = require('./lib/scheduler.js');
+const {
+  routineState, loadRoutineState, saveRoutineState, recordRoutineRun,
+  startScheduler, getNextRun, executeRoutine,
+} = schedulerLib;
 // Hand lib/agents its root-owned dependencies (see the module headers).
 // routineState goes over BY IDENTITY: the scheduler mutates it in place.
 agentsDiscovery.setRoutineState(routineState);
@@ -504,7 +505,7 @@ workspaceAnalysis.wireAnalysisDeps({ parseSkillFile });
 workspaceScaffold.wireScaffoldDeps({ invalidateAgentCache, rebaselineAgentsWatcher });
 const codexGlue = require('./lib/runtime/codex-glue.js');
 const {
-  getCodexAppServer, waitForCodexReady, shutdownCodexAppServer,
+  shutdownCodexAppServer,
   readAgentInstructions, wireCodexDelegate, startCodexTurn,
 } = codexGlue;
 codexGlue.wireCodexGlueDeps({
@@ -519,38 +520,12 @@ codexGlue.wireCodexGlueDeps({
   getPermissionTimeoutMs: () => PERMISSION_TIMEOUT_MS,
 });
 
-function loadRoutineState() {
-  for (const key of Object.keys(routineState)) delete routineState[key];
-  try {
-    const file = path.join(rundockDir(), 'routine-state.json');
-    const saved = JSON.parse(fs.readFileSync(file, 'utf-8'));
-    for (const [key, state] of Object.entries(saved)) {
-      if (!state || typeof state.lastRun !== 'string') continue;
-      // A run that was 'running' when the server died never finished; surface
-      // that honestly. lastRun stays, so the run still suppresses a re-fire
-      // (the work was started; firing it again is the bug this file prevents).
-      if (state.status === 'running') state.status = 'interrupted';
-      routineState[key] = state;
-    }
-  } catch (e) { /* missing or unreadable file: start empty */ }
-}
-
-function saveRoutineState() {
-  const dir = rundockDir();
-  fs.mkdirSync(dir, { recursive: true });
-  fs.writeFileSync(path.join(dir, 'routine-state.json'), JSON.stringify(routineState, null, 2));
-}
-
-function recordRoutineRun(key, state) {
-  routineState[key] = state;
-  try {
-    saveRoutineState();
-  } catch (e) {
-    // Persistence is protection for the NEXT process; this one already has
-    // the in-memory state. An unwritable .rundock must not kill the scheduler.
-    console.error('[Scheduler] Failed to persist routine state:', e && e.message ? e.message : e);
-  }
-}
+schedulerLib.wireSchedulerDeps({
+  // Root spawn plumbing until its own extraction slice; the client set is
+  // read through an accessor because wss is created later at boot.
+  spawnClaude, getBareArgs, modelArgs, getSpawnEnv,
+  getWssClients: () => wss.clients,
+});
 
 // ===== AGENT HELPERS =====
 
@@ -796,152 +771,8 @@ function watchOpenFile(ws, relPath, fullPath) {
 // parseCapabilities, parseRoutines, parsePrompts, parseSkills) and titleCase
 // live in lib/agents/discovery.js.
 
-// ===== SCHEDULER =====
-
-function startScheduler() {
-  const checkInterval = 60 * 1000; // Check every 60 seconds
-
-  setInterval(() => {
-    const agents = discoverAgents();
-    const now = new Date();
-
-    for (const agent of agents) {
-      if (!agent.routines) continue;
-      for (const routine of agent.routines) {
-        const key = `${agent.id}:${routine.name}`;
-        const nextRun = getNextRun(routine.schedule, routineState[key]?.lastRun);
-        if (nextRun && now >= nextRun) {
-          console.log(`[Scheduler] Running routine: ${routine.name} (${agent.name})`);
-          executeRoutine(agent, routine, key);
-        }
-      }
-    }
-  }, checkInterval).unref(); // see heartbeat unref note: listener keeps process alive in production
-}
-
-function getNextRun(schedule, lastRunISO) {
-  if (!schedule) return null;
-  const now = new Date();
-  const s = schedule.toLowerCase();
-
-  // Parse "every day at HH:MM"
-  const dailyMatch = s.match(/every day at (\d{2}):(\d{2})/);
-  if (dailyMatch) {
-    // Don't re-run if already ran today. This suppression (fed by the
-    // persisted routine state) is the ONLY thing standing between a due
-    // routine and a duplicate fire, which is why it is checked first.
-    if (lastRunISO) {
-      const lastRun = new Date(lastRunISO);
-      if (lastRun.toDateString() === now.toDateString() && lastRun.getHours() >= parseInt(dailyMatch[1])) return null;
-    }
-    const target = new Date(now);
-    target.setHours(parseInt(dailyMatch[1]), parseInt(dailyMatch[2]), 0, 0);
-    // A target already past today stays TODAY: the scheduler's `now >= nextRun`
-    // check fires it on the next tick (same-day catch-up). The previous code
-    // rolled it to tomorrow, which meant the fire condition was only
-    // satisfiable in the single millisecond HH:MM:00.000 - routines whose
-    // tick did not land exactly on that instant never fired at all.
-    return target;
-  }
-
-  // Parse "every [weekday] at HH:MM"
-  const weeklyMatch = s.match(/every (\w+) at (\d{2}):(\d{2})/);
-  if (weeklyMatch) {
-    const days = { sunday: 0, monday: 1, tuesday: 2, wednesday: 3, thursday: 4, friday: 5, saturday: 6 };
-    const targetDay = days[weeklyMatch[1]];
-    if (targetDay === undefined) return null;
-    // Suppression first, same reasoning as the daily branch.
-    if (lastRunISO) {
-      const lastRun = new Date(lastRunISO);
-      const daysSinceLastRun = (now - lastRun) / (1000 * 60 * 60 * 24);
-      if (daysSinceLastRun < 1 && lastRun.getDay() === targetDay) return null;
-    }
-    const target = new Date(now);
-    target.setHours(parseInt(weeklyMatch[2]), parseInt(weeklyMatch[3]), 0, 0);
-    const daysUntil = (targetDay - now.getDay() + 7) % 7;
-    target.setDate(target.getDate() + daysUntil);
-    // On the target weekday a past-due target stays TODAY so the scheduler
-    // fires it (same-day catch-up); the suppression above prevents re-fires.
-    // See the daily branch for why the old roll-forward meant never firing.
-    return target;
-  }
-
-  return null;
-}
-
-function executeRoutine(agent, routine, key) {
-  const startTime = Date.now();
-  // Persisted immediately: if the server dies mid-run, the restarted process
-  // still knows the run started and will not fire it again in the same window.
-  recordRoutineRun(key, { lastRun: new Date().toISOString(), status: 'running', duration: null });
-
-  // Notify connected clients
-  broadcastRoutineUpdate();
-
-  const recordOutcome = (ok) => {
-    const duration = Math.round((Date.now() - startTime) / 1000);
-    recordRoutineRun(key, {
-      lastRun: new Date().toISOString(),
-      status: ok ? 'completed' : 'failed',
-      duration
-    });
-    console.log(`[Scheduler] Routine "${routine.name}" ${ok ? 'completed' : 'failed'} (${duration}s)`);
-    recordEvent('routine_run', { agent: agent.id, runtime: agent.runtime || 'claude', d: { routine: routine.name, status: ok ? 'completed' : 'failed', duration } });
-    broadcastRoutineUpdate();
-  };
-
-  if (agent.runtime === 'codex') {
-    // Codex agents run their routines on the shared Codex app-server: one
-    // fresh thread per run, the routine prompt travelling with the agent's
-    // instructions (Codex has no --agent equivalent). Routines run
-    // unattended with nobody to approve escalations, so approvalPolicy is
-    // an EXPLICIT 'never' (the client refuses to default to it):
-    // sandbox-blocked actions fail instead of hanging on an approval,
-    // matching the retired exec mode. The agent's plan choice is honoured
-    // even for unattended work.
-    const routinePrompt = [readAgentInstructions(agent), buildSystemPrompt(agent), routine.prompt].filter(Boolean).join('\n\n');
-    (async () => {
-      const server = await getCodexAppServer();
-      await waitForCodexReady(server);
-      const { threadId } = await server.startThread({
-        cwd: WORKSPACE,
-        model: agent.model || undefined,
-        sandbox: 'workspace-write',
-        approvalPolicy: 'never',
-      });
-      const sub = server.startTurn(threadId, routinePrompt);
-      const status = await new Promise((resolve) => {
-        sub.on('event', (ev) => { if (ev.type === 'done') resolve(ev.status); });
-      });
-      return status === 'completed';
-    })().then(recordOutcome, (err) => {
-      console.error(`[Scheduler] Codex routine "${routine.name}" failed to run: ${err.message}`);
-      recordOutcome(false);
-    });
-    return;
-  }
-
-  // Routines run unattended (no user to approve), so bypass permissions.
-  const args = [...getBareArgs(), ...modelArgs(agent), '--print', '--output-format', 'stream-json', '--verbose', '--dangerously-skip-permissions'];
-  if (agent.id !== 'default') args.push('--agent', agent.id);
-  args.push(routine.prompt);
-
-  const proc = spawnClaude(args, {
-    cwd: WORKSPACE,
-    env: getSpawnEnv(null),
-    stdio: ['ignore', 'pipe', 'pipe']
-  });
-
-  proc.on('close', (code) => recordOutcome(code === 0));
-}
-
-function broadcastRoutineUpdate() {
-  const agents = discoverAgents();
-  const msg = JSON.stringify({ type: 'agents', agents });
-  wss.clients.forEach(client => {
-    if (client.readyState === 1) client.send(msg);
-  });
-}
+// The scheduler (startScheduler, getNextRun, executeRoutine, the routine
+// broadcast) lives in lib/scheduler.js.
 
 // Workspace analysis (Seven Signals, SKILL_CLUSTERS) lives in
 // lib/workspace/analysis.js.
